+++ src/assets/styles/main.css
@@ -14,8 +14,6 @@
     --font-roboto: 'Roboto Mono', sans-serif;
     --font-flex: 'Roboto Flex', sans-serif;
     --font-main: 'Bebas Neue', sans-serif;
-
-   
 }
 
 @layer base {
